# improve perf with array pre-allocate