--- a/.changeset/giant-schools-explode.md
+++ b/.changeset/giant-schools-explode.md
@@ -1,0 +1,5 @@
+---
+'preact-render-to-string': patch
+---
+
+Improve performance by pre-allocating arrays