```
Text:
baseline x 434.471 ops/sec ±2.09 (89 runs sampled)
current x 455.816 ops/sec ±1.35 (88 runs sampled)

Fastest is current

SearchResults:
baseline x 2,124.763 ops/sec ±3.50 (92 runs sampled)
current x 2,174.318 ops/sec ±4.29 (88 runs sampled)

Fastest is current

ColorPicker:
baseline x 12,934.781 ops/sec ±1.34 (95 runs sampled)
current x 13,121.543 ops/sec ±0.87 (97 runs sampled)

Fastest is current
```